See comments
Implemented opening. Fixed the scaling to fit in RGB for both Hough and Otsu. Hough now goes from -90 to 89 for theta and -diag to diag for r. 2*diag makes it hold pos and negative, then added a group factor currently set to 10 for grouping rhos together. Image produced by it looks very good. We'll need to multiple by group and subtract diag to figure out the real pos/neg of r, as will we subtract 90 from theta for now to get 0-180 to -90 to 90 and 0-r/10 to -r to r